tools/helm/requirements.txt
mcp>=1.0.0
PyYAML>=6.0
